feat(SDK): add support for start button
The Oculus SDK supports a start button as the left controller
menu button can be overriden to be used as a start button although
the right controller menu button is reserved for Oculus Home.

This adds the support for the start button to the Oculus SDK and to
any other VRTK classes or prefabs that deal with buttons.